backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-docx
python-pptx
google-generativeai
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
requests
psycopg2-binary